generator/requirements.txt
tensorflow-gpu==1.14
scipy==1.3.3
numpy==1.14.5
Pillow==6.2.1
tqdm==4.64.1
requests==2.22.0
dlib==19.24.0
opencv-python==4.6.0.66
torchvision==0.11.2
Pyro4==4.82